requirements.txt
# keep in alphabetical order
dm_control
omegaconf==2.3.0
Pillow
pin==3.7.0
pygame
pyspacemouse
PyQt6
pyquaternion
pyrealsense2
pyzmq
pure-python-adb
quaternion
setuptools
tyro
ur-rtde
zmq
xarm
xarm-python-sdk
numpy-quaternion
termcolor
matplotlib
natsort
mediapy
transforms3d
psutil